small changes regarding some TODOs
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@19245 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/ci/redundancyremover/RedundantRequirementsRemoverSignStateImplementation.java
@@ -76,9 +76,9 @@ RedundantRequirementsRemoverImplementation<SignState, SIGN>{
     // if pVarOrConst number, number<0 MINUS, number=0 ZERO, number>0 PLUS
     // otherwise getSignForVariable
 
-    int constant;
+    double constant;
     try {
-      constant = Integer.parseInt(pVarOrConst);
+      constant = Double.parseDouble(pVarOrConst);
       if (constant < 0) {
         return SIGN.MINUS;
       } else if (constant == 0) {
@@ -87,7 +87,7 @@ RedundantRequirementsRemoverImplementation<SignState, SIGN>{
         return SIGN.PLUS;
       }
     } catch (NumberFormatException e) {
-      // TODO
+      // pVarOrConst is var and handled in next return
     }
 
     return pAbstractState.getSignForVariable(pVarOrConst);
